feat(core:loan): migrated to cmp
