+++ core/common/src/androidMain/kotlin/org/mifos/mobile/core/common/FromatAmount.android.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.core.common
+
+import java.text.DecimalFormat
+
+actual fun formatAmount(amount: Double): String {
+    val formatter = DecimalFormat("#,##0.00") // Ensures two decimal places
+    return formatter.format(amount)
+}
