+++ core/common/src/jsMain/kotlin/org/mifos/mobile/core/common/FromatAmount.js.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.core.common
+
+actual fun formatAmount(amount: Double): String {
+    return amount.asDynamic().toFixed(2) as String
+}
