+++ core/common/src/nativeMain/kotlin/org/mifos/mobile/core/common/FromatAmount.native.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.core.common
+
+import platform.Foundation.NSNumber
+import platform.Foundation.NSNumberFormatter
+import platform.Foundation.NSNumberFormatterDecimalStyle
+
+actual fun formatAmount(amount: Double): String {
+    val formatter = NSNumberFormatter().apply {
+        numberStyle = NSNumberFormatterDecimalStyle
+        maximumFractionDigits = 2.toULong()
+        minimumFractionDigits = 2.toULong()
+    }
+    return formatter.stringFromNumber(NSNumber(amount)) ?: amount.toString()
+}
