+++ core/common/src/wasmJsMain/kotlin/org/mifos/mobile/core/common/FromatAmount.wasmJs.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2025 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.core.common
+
+@JsFun("a => a.toFixed(2)")
+external fun jsToFixed(amount: Double): String
+
+actual fun formatAmount(amount: Double): String {
+    return jsToFixed(amount)
+}
