+++ feature/loan/build.gradle.kts
@@ -8,14 +8,32 @@
  * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
  */
 plugins {
-    alias(libs.plugins.mifos.android.feature)
-    alias(libs.plugins.mifos.android.library.compose)
+//    alias(libs.plugins.mifos.android.feature)
+//    alias(libs.plugins.mifos.android.library.compose)
+    alias(libs.plugins.mifos.cmp.feature)
+    alias(libs.plugins.kotlin.serialization)
+//    id(libs.plugins.kotlin.parcelize.get().pluginId)
 }
 
 android {
     namespace = "org.mifos.mobile.feature.loan"
 }
 
-dependencies {
-    implementation(projects.core.qrcode)
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            // core modules
+            implementation(projects.core.qrcode)
+
+            implementation(compose.material3)
+            implementation(compose.foundation)
+            implementation(compose.ui)
+            implementation(compose.components.resources)
+            implementation(compose.components.uiToolingPreview)
+            implementation(libs.jb.kotlin.stdlib)
+            implementation(libs.kotlin.reflect)
+            implementation(libs.kotlinx.serialization.json)
+
+        }
+    }
 }
