+++ feature/loan/src/androidMain/kotlin/AndroidManifest.xml
