+++ feature/loan/src/commonMain/composeResources/drawable/ic_assignment_turned_in_black_24dp.xml
