+++ feature/loan/src/commonMain/composeResources/drawable/ic_charges.xml
