+++ feature/loan/src/commonMain/composeResources/drawable/ic_check_circle_green_24px.xml
