+++ feature/loan/src/commonMain/composeResources/drawable/ic_compare_arrows_black_24dp.xml
