+++ feature/loan/src/commonMain/composeResources/drawable/ic_edit_black_24dp.xml
