+++ feature/loan/src/commonMain/composeResources/drawable/ic_qrcode_scan.xml
