+++ feature/loan/src/commonMain/composeResources/drawable/ic_report_problem_red_24px.xml
