+++ feature/loan/src/commonMain/composeResources/drawable/ic_surveys_48px.xml
