+++ feature/loan/src/commonMain/composeResources/values/strings.xml
