+++ feature/loan/src/commonMain/kotlin/org/mifos/mobile/feature/loan/loanAccountSummary/LoanAccountSummaryScreen.kt
@@ -1,0 +1,299 @@
+/*
+ * Copyright 2024 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.loan.loanAccountSummary
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedCard
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import mifos_mobile.feature.loan.generated.resources.Res
+import mifos_mobile.feature.loan.generated.resources.account_short
+import mifos_mobile.feature.loan.generated.resources.account_status
+import mifos_mobile.feature.loan.generated.resources.active_uc
+import mifos_mobile.feature.loan.generated.resources.fees
+import mifos_mobile.feature.loan.generated.resources.fees_waived
+import mifos_mobile.feature.loan.generated.resources.ic_check_circle_green_24px
+import mifos_mobile.feature.loan.generated.resources.ic_report_problem_red_24px
+import mifos_mobile.feature.loan.generated.resources.inactive_uc
+import mifos_mobile.feature.loan.generated.resources.interest
+import mifos_mobile.feature.loan.generated.resources.interest_waived
+import mifos_mobile.feature.loan.generated.resources.loan_product
+import mifos_mobile.feature.loan.generated.resources.loan_summary
+import mifos_mobile.feature.loan.generated.resources.outstanding_balance
+import mifos_mobile.feature.loan.generated.resources.penalties
+import mifos_mobile.feature.loan.generated.resources.penalties_waived
+import mifos_mobile.feature.loan.generated.resources.principal
+import mifos_mobile.feature.loan.generated.resources.string_and_double
+import mifos_mobile.feature.loan.generated.resources.total_paid
+import mifos_mobile.feature.loan.generated.resources.total_repayment
+import org.jetbrains.compose.resources.stringResource
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.koin.compose.viewmodel.koinViewModel
+import org.mifos.mobile.core.designsystem.component.MifosCard
+import org.mifos.mobile.core.designsystem.component.MifosScaffold
+import org.mifos.mobile.core.designsystem.theme.MifosMobileTheme
+import org.mifos.mobile.core.model.entity.accounts.loan.LoanWithAssociations
+import org.mifos.mobile.core.ui.component.MifosErrorComponent
+import org.mifos.mobile.core.ui.component.MifosProgressIndicatorOverlay
+import org.mifos.mobile.core.ui.component.MifosTextTitleDescDrawableSingleLine
+import org.mifos.mobile.core.ui.component.MifosTextTitleDescSingleLine
+import org.mifos.mobile.core.ui.utils.EventsEffect
+
+@Composable
+internal fun LoanAccountSummaryScreen(
+    navigateBack: () -> Unit,
+    modifier: Modifier = Modifier,
+    viewModel: LoanAccountSummaryViewModel = koinViewModel(),
+) {
+    val state by viewModel.stateFlow.collectAsStateWithLifecycle()
+
+    EventsEffect(viewModel.eventFlow) { event ->
+        when (event) {
+            LoanAccountSummaryEvent.NavigateBack -> navigateBack.invoke()
+        }
+    }
+    LoanAccountSummaryScreen(
+        state = state,
+        onAction = remember(viewModel) {
+            { viewModel.trySendAction(it) }
+        },
+        modifier = modifier,
+    )
+}
+
+@Composable
+private fun LoanAccountSummaryScreen(
+    state: LoanAccountSummaryState,
+    modifier: Modifier = Modifier,
+    onAction: (LoanAccountSummaryAction) -> Unit,
+) {
+    MifosScaffold(
+        topBarTitle = stringResource(Res.string.loan_summary),
+        backPress = { (onAction(LoanAccountSummaryAction.BackPress)) },
+        content = {
+            Box(modifier = Modifier.padding(it)) {
+                when (state.dialogState) {
+                    is LoanAccountSummaryState.DialogState.Error -> MifosProgressIndicatorOverlay()
+                    LoanAccountSummaryState.DialogState.Loading ->
+                        MifosErrorComponent(isNetworkConnected = state.isOnline)
+                    null -> LoanAccountSummaryContent(
+                        loanWithAssociations = state.loanAccountAssociations,
+                        modifier = modifier,
+                    )
+                }
+            }
+        },
+    )
+}
+
+@Composable
+private fun LoanAccountSummaryContent(
+    loanWithAssociations: LoanWithAssociations?,
+    modifier: Modifier = Modifier,
+) {
+    var currencySymbol = loanWithAssociations?.currency?.displaySymbol
+    if (currencySymbol == null) {
+        currencySymbol = loanWithAssociations?.currency?.code ?: ""
+    }
+
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+    ) {
+        MifosTextTitleDescSingleLine(
+            modifier = Modifier.padding(horizontal = 14.dp),
+            title = stringResource(Res.string.account_short),
+            description = loanWithAssociations?.accountNo ?: "",
+        )
+
+        MifosTextTitleDescSingleLine(
+            modifier = Modifier.padding(horizontal = 14.dp),
+            title = stringResource(Res.string.loan_product),
+            description = loanWithAssociations?.loanProductName ?: "",
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        MifosCard(
+            modifier = Modifier.fillMaxWidth(),
+            colors = CardDefaults.outlinedCardColors(
+                containerColor = MaterialTheme.colorScheme.background,
+            ),
+        ) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(14.dp),
+            ) {
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.principal),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.principal ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.interest),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.interestCharged ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.fees),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.feeChargesCharged ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.penalties),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.penaltyChargesCharged ?: 0.0,
+                    ),
+                )
+            }
+        }
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        OutlinedCard(
+            modifier = Modifier.fillMaxWidth(),
+            colors = CardDefaults.outlinedCardColors(
+                containerColor = MaterialTheme.colorScheme.background,
+            ),
+        ) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(14.dp),
+            ) {
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.total_repayment),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.totalExpectedRepayment ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.total_paid),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.totalRepayment ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.interest_waived),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.interestWaived ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.penalties_waived),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.penaltyChargesWaived ?: 0.0,
+                    ),
+                )
+
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.fees_waived),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.feeChargesWaived ?: 0.0,
+                    ),
+                )
+            }
+        }
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        OutlinedCard(
+            modifier = Modifier.fillMaxWidth(),
+            colors = CardDefaults.outlinedCardColors(
+                containerColor = MaterialTheme.colorScheme.background,
+            ),
+        ) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(14.dp),
+            ) {
+                MifosTextTitleDescSingleLine(
+                    title = stringResource(Res.string.outstanding_balance),
+                    description = stringResource(
+                        Res.string.string_and_double,
+                        currencySymbol,
+                        loanWithAssociations?.summary?.totalOutstanding ?: 0.0,
+                    ),
+                )
+                MifosTextTitleDescDrawableSingleLine(
+                    title = stringResource(Res.string.account_status),
+                    description = if (loanWithAssociations?.status?.active == true) {
+                        stringResource(
+                            Res.string.active_uc,
+                        )
+                    } else {
+                        stringResource(Res.string.inactive_uc)
+                    },
+                    imageResId = if (loanWithAssociations?.status?.active == true) {
+                        Res.drawable.ic_check_circle_green_24px
+                    } else {
+                        Res.drawable.ic_report_problem_red_24px
+                    },
+                )
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun LoanAccountSummaryPreview() {
+    MifosMobileTheme {
+        LoanAccountSummaryScreen(
+            state = LoanAccountSummaryState(dialogState = null),
+            onAction = {},
+            modifier = Modifier,
+        )
+    }
+}
