+++ feature/loan/src/commonMain/kotlin/org/mifos/mobile/feature/loan/loanAccountTransaction/LoanAccountTransactionScreen.kt
@@ -1,0 +1,198 @@
+/*
+ * Copyright 2024 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.loan.loanAccountTransaction
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import mifos_mobile.feature.loan.generated.resources.Res
+import mifos_mobile.feature.loan.generated.resources.atm_icon
+import mifos_mobile.feature.loan.generated.resources.ic_local_atm_black_24dp
+import mifos_mobile.feature.loan.generated.resources.string_and_string
+import mifos_mobile.feature.loan.generated.resources.transactions
+import org.jetbrains.compose.resources.painterResource
+import org.jetbrains.compose.resources.stringResource
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.koin.compose.viewmodel.koinViewModel
+import org.mifos.mobile.core.common.CurrencyFormatter
+import org.mifos.mobile.core.common.Utils.formatTransactionType
+import org.mifos.mobile.core.designsystem.component.MifosTopBar
+import org.mifos.mobile.core.designsystem.theme.MifosMobileTheme
+import org.mifos.mobile.core.model.entity.Transaction
+import org.mifos.mobile.core.model.entity.accounts.loan.LoanWithAssociations
+import org.mifos.mobile.core.ui.component.MifosErrorComponent
+import org.mifos.mobile.core.ui.component.MifosProgressIndicatorOverlay
+import org.mifos.mobile.core.ui.utils.EventsEffect
+
+@Composable
+internal fun LoanAccountTransactionScreen(
+    navigateBack: () -> Unit,
+    modifier: Modifier = Modifier,
+    viewModel: LoanAccountTransactionViewModel = koinViewModel(),
+) {
+    val state by viewModel.stateFlow.collectAsStateWithLifecycle()
+
+    EventsEffect(viewModel.eventFlow) { event ->
+        when (event) {
+            LoanAccountTransactionEvent.NavigateBack -> navigateBack.invoke()
+        }
+    }
+
+    LoanAccountTransactionScreen(
+        state = state,
+        onAction = remember(viewModel) {
+            { viewModel.trySendAction(it) }
+        },
+        modifier = modifier,
+    )
+}
+
+@Composable
+private fun LoanAccountTransactionScreen(
+    state: LoanAccountTransactionState,
+    modifier: Modifier = Modifier,
+    onAction: (LoanAccountTransactionAction) -> Unit,
+) {
+    Column(modifier = modifier.fillMaxSize().padding(16.dp)) {
+        MifosTopBar(
+            backPress = { (onAction(LoanAccountTransactionAction.BackPress)) },
+            topBarTitle = stringResource(Res.string.transactions),
+        )
+
+        Box(modifier = Modifier.weight(1f)) {
+            when (state.dialogState) {
+                is LoanAccountTransactionState.DialogState.Error -> {
+                    MifosErrorComponent(
+                        isNetworkConnected = state.isOnline,
+                        isEmptyData = false,
+                    )
+                }
+                LoanAccountTransactionState.DialogState.Loading -> MifosProgressIndicatorOverlay()
+                null -> state.loanWithAssociations?.let { LoanAccountTransactionContent(loanWithAssociations = it) }
+            }
+        }
+    }
+}
+
+@Composable
+private fun LoanAccountTransactionContent(
+    loanWithAssociations: LoanWithAssociations,
+    modifier: Modifier = Modifier,
+) {
+    var currencySymbol = loanWithAssociations.currency?.displaySymbol
+    if (currencySymbol == null) {
+        currencySymbol = loanWithAssociations.currency?.code ?: ""
+    }
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+    ) {
+        Text(
+            text = loanWithAssociations.loanProductName ?: "",
+            style = MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.Bold),
+            color = MaterialTheme.colorScheme.onSurface,
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        LazyColumn {
+            items(items = loanWithAssociations.transactions?.toList().orEmpty()) {
+                LoanAccountTransactionListItem(it)
+            }
+        }
+    }
+}
+
+@Composable
+private fun LoanAccountTransactionListItem(
+    transaction: Transaction?,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier = modifier.padding(8.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Image(
+            painter = painterResource(Res.drawable.ic_local_atm_black_24dp),
+            contentDescription = stringResource(Res.string.atm_icon),
+            modifier = Modifier.size(39.dp),
+        )
+
+        Spacer(modifier = Modifier.width(8.dp))
+
+        Column(modifier = Modifier.weight(1f)) {
+            Text(
+                text = formatTransactionType(transaction?.type?.value),
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurface,
+            )
+
+            Row {
+                Text(
+                    text = stringResource(
+                        Res.string.string_and_string,
+                        transaction?.currency?.displaySymbol ?: transaction?.currency?.code ?: "",
+                        CurrencyFormatter.format(
+                            transaction?.amount ?: 0.0,
+                            "",
+                            5,
+                        ),
+                    ),
+                    style = MaterialTheme.typography.labelMedium,
+                    modifier = Modifier
+                        .weight(1f)
+                        .alpha(0.7f),
+                    color = MaterialTheme.colorScheme.onSurface,
+                )
+                Text(
+                    text = transaction?.submittedOnDate.toString(),
+                    style = MaterialTheme.typography.labelMedium,
+                    modifier = Modifier.alpha(0.7f),
+                    color = MaterialTheme.colorScheme.onSurface,
+                )
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun LoanAccountTransactionScreenPreview() {
+    MifosMobileTheme {
+        LoanAccountTransactionScreen(
+            state = LoanAccountTransactionState(dialogState = null),
+            modifier = Modifier,
+            onAction = {},
+        )
+    }
+}
