+++ feature/loan/src/commonMain/kotlin/org/mifos/mobile/feature/loan/loanAccountWithdraw/LoanAccountWithdrawScreen.kt
@@ -1,0 +1,191 @@
+/*
+ * Copyright 2024 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+package org.mifos.mobile.feature.loan.loanAccountWithdraw
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.RectangleShape
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import kotlinx.coroutines.launch
+import mifos_mobile.feature.loan.generated.resources.Res
+import mifos_mobile.feature.loan.generated.resources.account_number
+import mifos_mobile.feature.loan.generated.resources.client_name
+import mifos_mobile.feature.loan.generated.resources.withdraw_loan
+import mifos_mobile.feature.loan.generated.resources.withdraw_loan_reason
+import org.jetbrains.compose.resources.stringResource
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.koin.compose.viewmodel.koinViewModel
+import org.mifos.mobile.core.designsystem.component.MifosButton
+import org.mifos.mobile.core.designsystem.component.MifosTopBar
+import org.mifos.mobile.core.designsystem.theme.MifosMobileTheme
+import org.mifos.mobile.core.model.entity.accounts.loan.LoanWithAssociations
+import org.mifos.mobile.core.ui.component.MifosErrorComponent
+import org.mifos.mobile.core.ui.component.MifosProgressIndicator
+import org.mifos.mobile.core.ui.component.MifosTitleDescSingleLineEqual
+import org.mifos.mobile.core.ui.utils.EventsEffect
+
+@Composable
+internal fun LoanAccountWithdrawScreen(
+    navigateBack: () -> Unit,
+    modifier: Modifier = Modifier,
+    viewModel: LoanAccountWithdrawViewModel = koinViewModel(),
+) {
+    val state by viewModel.stateFlow.collectAsStateWithLifecycle()
+    val scope = rememberCoroutineScope()
+
+    val snackbarHostState = remember { SnackbarHostState() }
+
+    EventsEffect(viewModel.eventFlow) { event ->
+        when (event) {
+            LoanAccountWithdrawEvent.NavigateBack -> navigateBack.invoke()
+            is LoanAccountWithdrawEvent.ShowToast -> {
+                scope.launch {
+                    snackbarHostState.showSnackbar(event.message)
+                }
+            }
+        }
+    }
+
+    LoanAccountWithdrawScreen(
+        state = state,
+        onAction = remember(viewModel) {
+            { viewModel.trySendAction(it) }
+        },
+        modifier = modifier,
+    )
+}
+
+@Composable
+private fun LoanAccountWithdrawScreen(
+    state: LoanAccountWithdrawState,
+    onAction: (LoanAccountWithdrawAction) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(16.dp)
+            .background(MaterialTheme.colorScheme.background),
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+    ) {
+        MifosTopBar(
+            backPress = { onAction(LoanAccountWithdrawAction.BackPress) },
+            topBarTitle = stringResource(Res.string.withdraw_loan),
+        )
+
+        Spacer(modifier = Modifier.height(8.dp))
+
+        Box(modifier = Modifier.weight(1f)) {
+            LoanAccountWithdrawContent(
+                loanWithAssociations = state.loanWithAssociations,
+                state = state,
+                onAction = onAction,
+            )
+
+            when (state.dialogState) {
+                is LoanAccountWithdrawState.DialogState.Loading -> {
+                    MifosProgressIndicator(
+                        modifier = Modifier
+                            .fillMaxSize()
+                            .background(MaterialTheme.colorScheme.background.copy(alpha = 0.7f)),
+                    )
+                }
+
+                is LoanAccountWithdrawState.DialogState.Error -> {
+                    MifosErrorComponent(message = state.dialogState.message)
+                }
+
+                null -> Unit
+            }
+        }
+    }
+}
+
+@Composable
+private fun LoanAccountWithdrawContent(
+    loanWithAssociations: LoanWithAssociations?,
+    state: LoanAccountWithdrawState,
+    onAction: (LoanAccountWithdrawAction) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Column(
+        modifier = modifier
+            .padding(16.dp)
+            .fillMaxSize(),
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+    ) {
+        MifosTitleDescSingleLineEqual(
+            title = stringResource(Res.string.client_name),
+            description = loanWithAssociations?.clientName ?: "",
+        )
+
+        MifosTitleDescSingleLineEqual(
+            title = stringResource(Res.string.account_number),
+            description = loanWithAssociations?.accountNo ?: "",
+        )
+
+        TextField(
+            modifier = Modifier.fillMaxWidth(),
+            value = state.loanReason,
+            placeholder = {
+                Text(text = stringResource(Res.string.withdraw_loan_reason))
+            },
+            onValueChange = { onAction(LoanAccountWithdrawAction.LoanReasonChanged(it)) },
+            textStyle = MaterialTheme.typography.bodyLarge,
+            colors = TextFieldDefaults.colors(
+                focusedContainerColor = MaterialTheme.colorScheme.background,
+                unfocusedContainerColor = MaterialTheme.colorScheme.background,
+            ),
+        )
+
+        MifosButton(
+            modifier = Modifier.fillMaxWidth(),
+            shape = RectangleShape,
+            onClick = { onAction(LoanAccountWithdrawAction.WithDrawClicked) },
+            content = {
+                Text(
+                    modifier = Modifier.padding(6.dp),
+                    text = stringResource(Res.string.withdraw_loan),
+                    style = MaterialTheme.typography.titleMedium,
+                )
+            },
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun LoanAccountWithdrawPreview() {
+    MifosMobileTheme {
+        LoanAccountWithdrawScreen(
+            state = LoanAccountWithdrawState(dialogState = null),
+            modifier = Modifier,
+            onAction = {},
+        )
+    }
+}
