--- feature/loan/src/main/java/org/mifos/mobile/feature/loan/loanAccountApplication/LoanApplicationScreen.kt
@@ -1,280 +1,0 @@
-/*
- * Copyright 2024 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
- */
-package org.mifos.mobile.feature.loan.loanAccountApplication
-
-import android.content.Context
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.google.gson.Gson
-import org.mifos.mobile.core.common.Network
-import org.mifos.mobile.core.common.utils.DateHelper
-import org.mifos.mobile.core.designsystem.components.MifosTopBar
-import org.mifos.mobile.core.designsystem.theme.MifosMobileTheme
-import org.mifos.mobile.core.model.entity.payload.LoansPayload
-import org.mifos.mobile.core.model.enums.LoanState
-import org.mifos.mobile.core.ui.component.MifosErrorComponent
-import org.mifos.mobile.core.ui.component.MifosProgressIndicatorOverlay
-import org.mifos.mobile.core.ui.utils.DevicePreviews
-import org.mifos.mobile.feature.loan.R
-
-@Composable
-internal fun LoanApplicationScreen(
-    navigateBack: () -> Unit,
-    reviewNewLoanApplication: (
-        loanState: LoanState,
-        loansPayloadString: String,
-        loanId: Long?,
-        loanName: String,
-        accountNo: String,
-    ) -> Unit,
-    submitUpdateLoanApplication: (
-        loanState: LoanState,
-        loansPayloadString: String,
-        loanId: Long?,
-        loanName: String,
-        accountNo: String,
-    ) -> Unit,
-    modifier: Modifier = Modifier,
-    viewModel: LoanApplicationViewModel = hiltViewModel(),
-) {
-    val context = LocalContext.current
-
-    val uiState by viewModel.loanUiState.collectAsStateWithLifecycle()
-    val uiData by viewModel.loanApplicationScreenData.collectAsStateWithLifecycle()
-    val loanState by viewModel.loanState.collectAsStateWithLifecycle()
-
-    LaunchedEffect(key1 = loanState) {
-        viewModel.loadLoanApplicationTemplate(loanState)
-    }
-
-    LoanApplicationScreen(
-        uiState = uiState,
-        uiData = uiData,
-        navigateBack = navigateBack,
-        loanState = loanState,
-        onRetry = { viewModel.loadLoanApplicationTemplate(loanState) },
-        modifier = modifier,
-        selectProduct = viewModel::productSelected,
-        selectPurpose = viewModel::purposeSelected,
-        setDisbursementDate = viewModel::setDisburseDate,
-        reviewClicked = {
-            viewModel.setPrincipalAmount(it)
-            getLoanPayload(
-                context = context,
-                loanState = loanState,
-                reviewNewLoanApplication = reviewNewLoanApplication,
-                submitUpdateLoanApplication = submitUpdateLoanApplication,
-                viewModel = viewModel,
-            )
-        },
-    )
-}
-
-@Composable
-private fun LoanApplicationScreen(
-    uiState: LoanApplicationUiState,
-    loanState: LoanState,
-    uiData: LoanApplicationScreenData,
-    navigateBack: () -> Unit,
-    selectProduct: (Int) -> Unit,
-    selectPurpose: (Int) -> Unit,
-    setDisbursementDate: (String) -> Unit,
-    reviewClicked: (String) -> Unit,
-    onRetry: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-
-    Scaffold(
-        modifier = modifier,
-        topBar = {
-            MifosTopBar(
-                modifier = Modifier.fillMaxWidth(),
-                navigateBack = { navigateBack() },
-                title = {
-                    Text(
-                        text = stringResource(
-                            id = if (loanState == LoanState.CREATE) {
-                                R.string.apply_for_loan
-                            } else {
-                                R.string.update_loan
-                            },
-                        ),
-                    )
-                },
-            )
-        },
-        content = {
-            Column(
-                modifier = Modifier
-                    .padding(it)
-                    .fillMaxSize(),
-            ) {
-                Box(modifier = Modifier.weight(1f)) {
-                    LoanApplicationContent(
-                        uiData = uiData,
-                        selectProduct = selectProduct,
-                        selectPurpose = selectPurpose,
-                        reviewClicked = reviewClicked,
-                        setDisbursementDate = setDisbursementDate,
-                    )
-                    when (uiState) {
-                        is LoanApplicationUiState.Success -> Unit
-
-                        is LoanApplicationUiState.Loading -> {
-                            MifosProgressIndicatorOverlay()
-                        }
-
-                        is LoanApplicationUiState.Error -> {
-                            MifosErrorComponent(
-                                isNetworkConnected = Network.isConnected(context),
-                                isEmptyData = false,
-                                isRetryEnabled = true,
-                                onRetry = onRetry,
-                            )
-                        }
-                    }
-                }
-            }
-        },
-    )
-}
-
-internal class UiStatesParameterProvider : PreviewParameterProvider<LoanApplicationUiState> {
-    override val values: Sequence<LoanApplicationUiState>
-        get() = sequenceOf(
-            LoanApplicationUiState.Error(R.string.something_went_wrong),
-            LoanApplicationUiState.Loading,
-            LoanApplicationUiState.Success,
-        )
-}
-
-@DevicePreviews
-@Composable
-private fun ReviewLoanApplicationScreenPreview(
-    @PreviewParameter(UiStatesParameterProvider::class)
-    loanApplicationUiState: LoanApplicationUiState,
-) {
-    MifosMobileTheme {
-        LoanApplicationScreen(
-            uiState = loanApplicationUiState,
-            uiData = LoanApplicationScreenData(),
-            loanState = LoanState.CREATE,
-            navigateBack = {},
-            selectPurpose = {},
-            selectProduct = {},
-            reviewClicked = {},
-            setDisbursementDate = {},
-            onRetry = {},
-        )
-    }
-}
-
-private fun getLoanPayload(
-    context: Context,
-    viewModel: LoanApplicationViewModel,
-    loanState: LoanState,
-    reviewNewLoanApplication: (
-        loanState: LoanState,
-        loansPayloadString: String,
-        loanId: Long?,
-        loanName: String,
-        accountNo: String,
-    ) -> Unit,
-    submitUpdateLoanApplication: (
-        loanState: LoanState,
-        loansPayloadString: String,
-        loanId: Long?,
-        loanName: String,
-        accountNo: String,
-    ) -> Unit,
-) {
-    val payload = LoansPayload().apply {
-        clientId = viewModel.loanTemplate.clientId.takeIf { loanState == LoanState.CREATE }
-        loanPurpose =
-            viewModel.loanApplicationScreenData.value.selectedLoanPurpose ?: "Not provided"
-        productName = viewModel.loanApplicationScreenData.value.selectedLoanProduct
-        currency = viewModel.loanApplicationScreenData.value.currencyLabel
-        if (viewModel.purposeId > 0) loanPurposeId = viewModel.purposeId
-        productId = viewModel.productId
-        principal =
-            viewModel.loanApplicationScreenData.value.principalAmount?.toDoubleOrNull() ?: 0.0
-        loanTermFrequency = viewModel.loanTemplate.termFrequency
-        loanTermFrequencyType = viewModel.loanTemplate.interestRateFrequencyType?.id
-        loanType = "individual".takeIf { loanState == LoanState.CREATE }
-        numberOfRepayments = viewModel.loanTemplate.numberOfRepayments
-        repaymentEvery = viewModel.loanTemplate.repaymentEvery
-        repaymentFrequencyType = viewModel.loanTemplate.interestRateFrequencyType?.id
-        interestRatePerPeriod = viewModel.loanTemplate.interestRatePerPeriod
-        expectedDisbursementDate = DateHelper.getSpecificFormat(
-            DateHelper.FORMAT_MMMM,
-            viewModel.loanApplicationScreenData.value.disbursementDate,
-        )
-        submittedOnDate = DateHelper.getSpecificFormat(
-            DateHelper.FORMAT_MMMM,
-            viewModel.loanApplicationScreenData.value.submittedDate,
-        ).takeIf { loanState == LoanState.CREATE }
-
-        transactionProcessingStrategyId = viewModel.loanTemplate.transactionProcessingStrategyId
-        amortizationType = viewModel.loanTemplate.amortizationType?.id
-        interestCalculationPeriodType = viewModel.loanTemplate.interestCalculationPeriodType?.id
-        interestType = viewModel.loanTemplate.interestType?.id
-    }
-
-    val loansPayloadString = Gson().toJson(payload)
-    when (loanState) {
-        LoanState.CREATE -> reviewNewLoanApplication(
-            loanState,
-            loansPayloadString,
-            viewModel.loanId.value,
-            context.getString(
-                R.string.string_and_string,
-                context.getString(R.string.new_loan_application) + " ",
-                viewModel.loanApplicationScreenData.value.clientName ?: "",
-            ),
-            context.getString(
-                R.string.string_and_string,
-                context.getString(R.string.account_number) + " ",
-                viewModel.loanApplicationScreenData.value.accountNumber ?: "",
-            ),
-        )
-
-        LoanState.UPDATE -> submitUpdateLoanApplication(
-            loanState,
-            loansPayloadString,
-            null,
-            context.getString(
-                R.string.string_and_string,
-                context.getString(R.string.update_loan_application) + " ",
-                viewModel.loanApplicationScreenData.value.clientName ?: "",
-            ),
-            context.getString(
-                R.string.string_and_string,
-                context.getString(R.string.account_number) + " ",
-                viewModel.loanApplicationScreenData.value.accountNumber ?: "",
-            ),
-        )
-    }
-}
