+++ settings.gradle.kts
@@ -57,7 +57,7 @@ include(":core:qrcode")
 //include(":core:testing")
 
 // Feature Modules
-//include(":feature:loan")
+include(":feature:loan")
 //include(":feature:beneficiary")
 //include(":feature:savings")
 //include(":feature:guarantor")
